Removed interactive mode from eth client and moved to separate subproject
The subproject is not currently on the build path
