+++ eth-ncurses-cli/CMakeLists.txt
@@ -16,17 +16,20 @@ if (${TARGET_PLATFORM} STREQUAL "w64")
 	target_link_libraries(eth shlwapi)
 	target_link_libraries(eth iphlpapi)
 	target_link_libraries(eth cryptopp)
+	target_link_libraries(eth ncurses)
 	target_link_libraries(eth form)
 	target_link_libraries(eth boost_system-mt-s)
 	target_link_libraries(eth boost_filesystem-mt-s)
 	target_link_libraries(eth boost_thread_win32-mt-s)
 	set(CMAKE_SHARED_LIBRARY_LINK_CXX_FLAGS)
 elseif (UNIX)
+	target_link_libraries(eth ncurses)
 	target_link_libraries(eth form)
 else ()
 	target_link_libraries(eth ${CRYPTOPP_LIBRARIES})
 	target_link_libraries(eth boost_system)
 	target_link_libraries(eth boost_filesystem)
+	target_link_libraries(eth ncurses)
 	target_link_libraries(eth form)
 	find_package(Threads REQUIRED)
 	target_link_libraries(eth ${CMAKE_THREAD_LIBS_INIT})
